server/issues: add API to list recommended issues to pledge towards
